Added support for Lingarr subtitles translation
