+++ bazarr/api/subtitles/subtitles.py
@@ -11,7 +11,7 @@ from languages.get_languages import alpha3_from_alpha2
 from utilities.path_mappings import path_mappings
 from utilities.video_analyzer import subtitles_sync_references
 from subtitles.tools.subsyncer import SubSyncer
-from subtitles.tools.translate import translate_subtitles_file
+from subtitles.tools.translate.main import translate_subtitles_file
 from subtitles.tools.mods import subtitles_apply_mods
 from subtitles.indexer.series import store_subtitles
 from subtitles.indexer.movies import store_subtitles_movie
@@ -19,7 +19,6 @@ from app.config import settings, empty_values
 from app.event_handler import event_stream
 
 from ..utils import authenticate
-
 
 api_ns_subtitles = Namespace('Subtitles', description='Apply mods/tools on external subtitles')
