+++ frontend/src/pages/Settings/Subtitles/index.tsx
@@ -546,6 +546,16 @@ const SettingsSubtitlesView: FunctionComponent = () => {
             You can generate it here: https://aistudio.google.com/apikey
           </Message>
         </CollapseBox>
+        <CollapseBox
+          settingKey="settings-translator-translator_type"
+          on={(val) => val === "lingarr"}
+        >
+          <Text
+            label="Lingarr endpoint"
+            settingKey="settings-translator-lingarr_url"
+          />
+          <Message>Base URL of Lingarr (e.g., http://localhost:9876)</Message>
+        </CollapseBox>
         <Check
           label="Add translation info at the beginning"
           settingKey="settings-translator-translator_info"
